[opt] recode logger to ts
