+++ package.json
@@ -2,7 +2,7 @@
     "name": "fntv",
     "version": "1.7.2",
     "description": "A fntv app built with Electron",
-    "main": "src/main/main.js",
+    "main": "dest/main/main.js",
     "scripts": {
         "start": "electron .",
         "build": "electron-builder"
