+++ src/public/constants.ts
@@ -1,0 +1,5 @@
+import * as fs from 'node:fs';
+import * as path from 'node:path';
+import { app } from 'electron';
+
+export const USER_DATA_PATH: string = path.join(app.getPath('home'), '.fntv');
